+++ drift/lib/src/query_builder/statements/insert.dart
@@ -301,6 +301,13 @@ extension InsertStatementWithDatabase<
   /// violations. If you have further uniqueness constraints, please use the
   /// general [insert] method with a [DoUpdate] including those columns in its
   /// [DoUpdate.target].
+  ///
+  /// Like the other [insert] methods, this returns the rowid of the last
+  /// insert. When the insert was turned into an update due to a conflicting
+  /// row, this behavior is somewhat confusing because the id of an (likely
+  /// unrelated) insert is returned instead.
+  /// Use [insertReturning] with an `onConflict` clause of
+  /// `DoUpdate((_) => entity)` to reliably get the written entity instead.
   Future<int> insertOnConflictUpdate(Insertable<Row> entity) {
     return insert(entity, onConflict: DoUpdate((_) => entity));
   }