+++ drift/lib/src/runtime/streams/store_impl.dart
@@ -122,7 +122,7 @@ base class LocalStreamQueryStore implements StreamQueryStore {
     _isShuttingDown = true;
 
     for (final stream in _activeKeyStreams.values) {
-      stream.close();
+      await stream.close();
     }
     // awaiting this is fine - the stream is never exposed to users and we don't
     // pause any subscriptions on it.
@@ -155,7 +155,7 @@ class QueryStream<Rows extends Object> {
       final queryListener = _QueryStreamListener(listener);
 
       if (_isClosed) {
-        listener.closeSync();
+        listener.close();
         return;
       }
 
@@ -300,12 +300,13 @@ class QueryStream<Rows extends Object> {
     }
   }
 
-  void close() {
+  Future<void> close() async {
     _isClosed = true;
-    for (final listener in _listeners) {
-      listener.controller.close();
-    }
+
+    final listenersDone = Future.wait(
+        [for (final listener in _listeners) listener.controller.close()]);
     _listeners.clear();
+    await listenersDone;
   }
 }