+++ drift/test/connections/isolate_test.dart
@@ -274,6 +274,7 @@ void main() {
         );
       });
 
+      stream.cancel();
       await db.close();
     }