+++ drift_flutter/lib/src/web.dart
@@ -18,6 +18,7 @@ DriftConnection driftDatabase({
       databaseName: name,
       sqlite3Uri: web.sqlite3Wasm,
       driftWorkerUri: web.driftWorker,
+      initializeDatabase: web.initializeDatabase,
     );
 
     (web.onResult ?? _defaultResultHandler)(result);