+++ sqlparser/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.41.2
+
+- Improve error message on unknown columns when it looks like the identifier
+  should have been a string literal.
+- Infer the output of `lag`, `lead` and `nth_value` to be nullable.
+
 ## 0.41.1
 
 - Support new features introduced in SQLite version 3.50.0.